Requested changes
reverted GP code
renamed CraftedItemCache to CraftedCacheManager
moved craftitem function into CraftedCacheManager
Modified MakeProductHandler for use the CraftedCacheManager if serverproperties on CRAFTING_MEMORY_WORK set to true
